Renamed the testcases
